refactor: decouple FireTable and move to common
Extracted `categorize` and `fromAccountSpecification` from `FireTable` into
standalone server-side functions/factories. Moved `Table` and `FireTable`
to `src/common/table/`. Updated all imports.
